requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
Django==5.2.5
django-bootstrap4==25.2
django-stdimage==6.0.2
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
PyMySQL==1.1.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
